Email to sellers when the purchase is made. Reduce stock in Publication table and more functions. Chat finally

+++ E-Commerce_Models/Compras.cs
@@ -17,7 +17,7 @@ namespace E_Commerce_Models
 
       // public int CompraNumeroRegistro { get; set; } // Ya que en una misma compra, puede venir dadas de varias publicaciones.
 
-       public float PrecioUnitarioCompra {  get; set; } 
+       public decimal PrecioUnitarioCompra {  get; set; } 
 
 
         public Compras(int idComprador, int IdPublicacion_aux , int Qcomprada_aux)
